Renamed casem resource containers
