+++ src/main/java/fi/otavanopisto/kuntaapi/server/integrations/casem/resources/CaseMPageContentResourceContainer.java
@@ -1,11 +1,11 @@
-package fi.otavanopisto.kuntaapi.server.integrations.casem.cache;
+package fi.otavanopisto.kuntaapi.server.integrations.casem.resources;
 
 import javax.enterprise.context.ApplicationScoped;
 
 import fi.otavanopisto.kuntaapi.server.resources.AbstractPageContentCache;
 
 @ApplicationScoped
-public class CaseMPageContentCache extends AbstractPageContentCache {
+public class CaseMPageContentResourceContainer extends AbstractPageContentCache {
 
   private static final long serialVersionUID = 5074106733901932323L;
