+++ src/main/java/fi/otavanopisto/kuntaapi/server/integrations/casem/resources/CaseMPageResourceContainer.java
@@ -1,11 +1,11 @@
-package fi.otavanopisto.kuntaapi.server.integrations.casem.cache;
+package fi.otavanopisto.kuntaapi.server.integrations.casem.resources;
 
 import javax.enterprise.context.ApplicationScoped;
 
 import fi.otavanopisto.kuntaapi.server.resources.AbstractPageResourceContainer;
 
 @ApplicationScoped
-public class CaseMPageCache extends AbstractPageResourceContainer {
+public class CaseMPageResourceContainer extends AbstractPageResourceContainer {
 
   private static final long serialVersionUID = -5748105617264972537L;
