+++ src/main/java/fi/otavanopisto/kuntaapi/server/integrations/casem/resources/CasemFileResourceContainer.java
@@ -1,11 +1,11 @@
-package fi.otavanopisto.kuntaapi.server.integrations.casem.cache;
+package fi.otavanopisto.kuntaapi.server.integrations.casem.resources;
 
 import javax.enterprise.context.ApplicationScoped;
 
 import fi.otavanopisto.kuntaapi.server.resources.AbstractFileResourceContainer;
 
 @ApplicationScoped
-public class CasemFileCache extends AbstractFileResourceContainer {
+public class CasemFileResourceContainer extends AbstractFileResourceContainer {
 
   private static final long serialVersionUID = 4993959656647694079L;
